+++ app/page.tsx
@@ -12,7 +12,7 @@ import BudgetCard from "@/components/BudgetCard";
 import {useCategoryBudgets_currentMonth} from "@/lib/firebase";
 import {CategoryBudget} from "@/lib/Interfaces";
 import LoadingAtAGlance from "@/components/layouts/LoadingAtAGlance";
-import {CustomButtons} from "@/components/CustomButtons";
+import AddExpensePopover from "@/components/AddExpensePopover";
 
 const PRIMARY_COL_HEIGHT = rem(400);
 
@@ -33,8 +33,8 @@ export default function Home() {
     return (
         <>
             <ThreeColumnLayout
-                one={<CustomButtons/>}
-                two={<Actions/>}
+                one={<AddExpensePopover/>}
+                two={<></>}
                 three={<AtAGlance
                     // userData={userData}
                     // user={user}
